+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * Allow setting `queues` to `false` to skip binding the default `messages` queue
+ * Add option `delay[daily_delay_queues]` in the transport definition
 
 7.3
 ---